requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
python-dotenv==1.0.1
pyarrow==16.1.0
